requirements.txt
streamlit
Pillow
imagehash
pyautogui
